refactor: extract select color options
